improve baseline tests to be more practical
- don't run them unless it's a release build
- make the existence of atomic tests more obvious with `make`
- share code between both baseline variants
- some output to learn about progress
