+++ Makefile
@@ -6,6 +6,9 @@ CARGO = $(shell command -v cargo)
 
 ##@ Development
 
+baseline-atomic: ## run very slow tests that single-step through all commits
+	RUST_BACKTRACE=1 cargo test --test baseline_atomic --release -- --nocapture
+
 test: ## run all tests with cargo
 	RUST_BACKTRACE=1 cargo test --test crates-index-diff
 	RUST_BACKTRACE=1 cargo test --test baseline --release
